fix: Bugs in UIVisualizerPanel; feat_end: Script fields in UIElementEditorPanel; feat_begin: Scene run

+++ Magma/src/Magma/UI/UIElement.cpp
@@ -40,7 +40,7 @@ UIElement& UIElement::SetSize(uint32_t width, uint32_t height) {
 	this->Height = height;
 	return *this;
 }
-UIElement& UIElement::SetPosition(uint32_t x, uint32_t y) {
+UIElement& UIElement::SetPosition(float x, float y) {
 	this->x = x;
 	this->y = y;
 	return *this;
